==> 0_predict_tf-idf.ipynb <==
import sys
import numpy as np
import pandas as pd
import jieba_fast
import jieba_fast.analyse


def tfidf(content:str) -> float:
    """
    Find TF-IDF value of the most valuable word in a single tweet
    """
    keyword = jieba_fast.analyse.extract_tags(content, topK=1, withWeight=True)
    if len(keyword) != 0:
        return keyword[0][1]
    else:
        return 0


df_tfidf = pd.read_csv('data_predict.txt',index_col=[0],header=0)
df_tfidf.head(2)

df_tfidf['tfidf'] = df_tfidf.text.apply(lambda x: tfidf(str(x)))
df_tfidf.to_csv("./data_predict_tfidf.txt",header=True)
df_tfidf.head(2)
