==> 5_random_forest.ipynb <==
# 暂时不用

import pandas as pd 
import numpy as np
# from sklearn import tree
from sklearn.ensemble import RandomForestRegressor
from sklearn.externals import joblib
from sklearn.metrics import mean_squared_error as mse

data_train=pd.read_csv('data_train.txt',index_col=[0],header=0)
data_valid=pd.read_csv('data_valid.txt',index_col=[0],header=0)
# 完整的
data_train = data_train.append(data_valid)

features_list = list(data_train.columns)[3:7] + list(data_train.columns)[8:]
train_subset=data_train.loc[:,features_list]
valid_subset=data_valid.loc[:,features_list]
train_subset.head()

def random_forest(data_train, data_test, estimator, depth, leaf_num, split_num, init=False):
    if init:
        data_test['repost_hat']=0
        data_test['comments_hat']=0
        data_test['likes_hat']=0
    else:
        pass
    
    #噪音数据预测值均为0,因此仅处理非噪音数据
    train=data_train[data_train['logit']==0]
    test=data_test[data_test['logit']==0]
    
    #分别对转发，评论和点赞建立三个森林
    forest_repost = RandomForestRegressor(n_estimators = estimator, oob_score = True, n_jobs = -1,random_state =50,
                                max_features = "auto", max_depth = depth, min_samples_leaf = leaf_num, min_samples_split = split_num) # criterion defalut by 'mse'
    forest_comments= RandomForestRegressor(n_estimators = estimator, oob_score = True, n_jobs = -1,random_state =50,
                                max_features = "auto", max_depth = depth, min_samples_leaf = leaf_num, min_samples_split = split_num)
    forest_likes = RandomForestRegressor(n_estimators = estimator, oob_score = True, n_jobs = -1,random_state =50,
                                max_features = "auto", max_depth = depth, min_samples_leaf = leaf_num, min_samples_split = split_num)
    #forest_repost = RandomForestRegressor(n_estimators = estimator, oob_score = True, n_jobs = -1
    #                            max_features = "auto") # criterion defalut by 'mse'
    #forest_comments= RandomForestRegressor(n_estimators = estimator, oob_score = True, n_jobs = -1,random_state =42,
    #                            max_features = "auto")
    #forest_likes = RandomForestRegressor(n_estimators = estimator, oob_score = True, n_jobs = -1,random_state =42,
    #                            max_features = "auto")
    
    #拟合三个森林
    regressor_train=train.drop(['repost','comments','likes','logit'],axis=1, inplace=False)
    repost_train=train.loc[:,['repost']]
    comments_train=train.loc[:,['comments']]
    likes_train=train.loc[:,['likes']]
    predict_repost=forest_repost.fit(regressor_train,repost_train.values.ravel()) # shape warning remove
    predict_comments=forest_comments.fit(regressor_train,comments_train.values.ravel())
    predict_likes=forest_likes.fit(regressor_train,likes_train.values.ravel())
    print(predict_repost.oob_score_, predict_comments.oob_score_, predict_likes.oob_score_)
    
    #预测测试集的数据
    regressor_test=test.drop(['repost','comments','likes','logit','repost_hat','comments_hat','likes_hat'],axis=1, inplace=False)
    repost_hat=np.round(predict_repost.predict(regressor_test),0)  #round函数只是返回四舍五入值，是浮点类型
    comments_hat=np.round(predict_comments.predict(regressor_test),0)
    likes_hat=np.round(predict_likes.predict(regressor_test),0)
    
    #将预测值赋值并设置为整数值
    data_test['repost_hat'][data_test['logit']==0]=repost_hat
    data_test['comments_hat'][data_test['logit']==0]=comments_hat
    data_test['likes_hat'][data_test['logit']==0]=likes_hat
    data_test['repost_hat']=data_test['repost_hat'].apply(lambda x:int(x))
    data_test['comments_hat']=data_test['comments_hat'].apply(lambda x:int(x))
    data_test['likes_hat']=data_test['likes_hat'].apply(lambda x:int(x))
    return data_test    


def precision(data):
    data['deviation_repost']=list(map(lambda x, y: abs(x-y)/(y+5), data['repost_hat'],data['repost']))
    #print (data['deviation_repost'])
    data['deviation_likes']=list(map(lambda x, y: abs(x-y)/(y+3), data['likes_hat'],data['likes']))
    #print (data['deviation_likes'])
    data['deviation_comments']=list(map(lambda x, y: abs(x-y)/(y+3), data['comments_hat'],data['comments']))
    #print (data['deviation_comments'])
    data['lcf_sum']=data['repost']+data['likes']+data['comments']
    #print (data['lcf_sum'])
    data['lcf_sum']=data['lcf_sum'].apply(lambda x: 100 if x>100 else x)
    data['precision_1_-0.8']=1-0.5*data['deviation_repost']-0.25*data['deviation_likes']-0.25*data['deviation_comments']-0.8
    #print (data['precision_1_-0.8'])
    data.loc[data['precision_1_-0.8']<=0,'sgn']=0
    data.loc[data['precision_1_-0.8']>0,'sgn']=1
    #print (data['sgn'])
    precision_=sum((data['lcf_sum']+1)*data['sgn'])/sum(data['lcf_sum']+1)
    return precision_

valid_subset=random_forest(train_subset, valid_subset, 500, 50, 10, 10) # estimator, depth, leaf_num, split_num, 
precision(valid_subset)

# 训练随机森林模型

### constants
estimators = 1000
oob = True
random_num = 42
jobs = -1
depth = 50
leaf_num = 10
split_num = 50

def precision(data):
    data['deviation_repost']=list(map(lambda x, y: abs(x-y)/(y+5), data['repost_hat'],data['repost']))
    #print (data['deviation_repost'])
    data['deviation_likes']=list(map(lambda x, y: abs(x-y)/(y+3), data['likes_hat'],data['likes']))
    #print (data['deviation_likes'])
    data['deviation_comments']=list(map(lambda x, y: abs(x-y)/(y+3), data['comments_hat'],data['comments']))
    #print (data['deviation_comments'])
    data['lcf_sum']=data['repost']+data['likes']+data['comments']
    #print (data['lcf_sum'])
    data['lcf_sum']=data['lcf_sum'].apply(lambda x: 100 if x>100 else x)
    data['precision_1_-0.8']=1-0.5*data['deviation_repost']-0.25*data['deviation_likes']-0.25*data['deviation_comments']-0.8
    #print (data['precision_1_-0.8'])
    data.loc[data['precision_1_-0.8']<=0,'sgn']=0
    data.loc[data['precision_1_-0.8']>0,'sgn']=1
    #print (data['sgn'])
    precision_=sum((data['lcf_sum']+1)*data['sgn'])/sum(data['lcf_sum']+1)
    return precision_


x=data_train[features_list]
y=data_train['repost']
model_repost =  RandomForestRegressor( n_estimators = estimators , oob_score = oob, random_state = random_num,  n_jobs = jobs,
                                max_features = "auto", max_depth = depth, min_samples_leaf = leaf_num, min_samples_split = split_num)
model_repost.fit(x,y)
#model_repost.score

x=data_train[features_list]
y=data_train['comments']
model_comments =  RandomForestRegressor( n_estimators = estimators , oob_score = oob, random_state = random_num,  n_jobs = jobs,
                                max_features = "auto", max_depth = depth, min_samples_leaf = leaf_num, min_samples_split = split_num)
model_comments.fit(x,y)
#model_comments.score

x=data_train[features_list]
y=data_train['likes']
model_likes =  RandomForestRegressor( n_estimators = estimators , oob_score = oob, random_state = random_num,  n_jobs = jobs,
                                max_features = "auto", max_depth = depth, min_samples_leaf = leaf_num, min_samples_split = split_num)
model_likes.fit(x,y)
#model_likes.score

model_repost.oob_score_, model_comments.oob_score_, model_likes.oob_score_

# Load Pre-trained Random Forest Model

### LOAD THE MODEL FROM THE DISK
model_repost = joblib.load('./model/model_repost.sav')
model_comments = joblib.load('./model/model_comments.sav')
model_likes = joblib.load('./model/model_likes.sav')

# 随机森林、逻辑回归、均值复合预测

import pandas as pd
import numpy as np
from sklearn import tree
from sklearn.ensemble import RandomForestRegressor

data_train=pd.read_csv('data_train.txt',index_col=[0],header=0)
data_valid=pd.read_csv('data_valid.txt',index_col=[0],header=0)
#data_train=data_train.append(data_valid)

features_list = ['tfidf', 'time_weekday', 'time_weekend', 'time_hour', 'panduan',
                   'length_all', 'length_chinese', 'english', 'non_ch', 'sharing', 'auto',
                   'interaction', 'book', 'mention', 'vote', 'lottery', 'emoji', 'video',
                   'http', 'stock', 'app', 'title', 'ad', 'hotwords', 'keywords',
                   'is_noise', 'number_in_train', 'forward_max', 'comment_max', 'like_max',
                   'forward_min', 'comment_min', 'like_min', 'forward_mean',
                   'comment_mean', 'like_mean', 'forward_more_ave_pr',
                   'comment_more_ave_pr', 'like_more_ave_pr', 'max_f/l', 'max_c/l',
                   'min_f/l', 'min_c/l', 'mean_f/l', 'mean_c/l', 'night', 'logit']
train_subset=data_train.loc[:,features_list]
valid_subset=data_valid.loc[:,features_list]

noise_cancellation = True


data_valid_test = data_valid
data_valid_test['repost_hat']   =0
data_valid_test['comments_hat'] =0
data_valid_test['likes_hat']    =0
if noise_cancellation == True:
    #噪音数据预测值均为0,因此仅处理非噪音数据[['logit']==0]
    x = data_valid_test[data_valid_test['logit'] == 0][features_list]
    data_valid_test[data_valid_test['logit'] == 0]['repost_hat'],   \
    data_valid_test[data_valid_test['logit'] == 0]['comments_hat'], \
    data_valid_test[data_valid_test['logit'] == 0]['likes_hat']     \
                                =                                   \
                                model_repost.predict(x),            \
                                model_comments.predict(x),          \
                                model_likes.predict(x)

else:
    x = data_valid_test[features_list]
    data_valid_test['repost_hat'],   \
    data_valid_test['comments_hat'], \
    data_valid_test['likes_hat']     \
                                =                                   \
                                model_repost.predict(x),            \
                                model_comments.predict(x),          \
                                model_likes.predict(x)
precision(data_valid_test[['repost', 'comments', 'likes', 'repost_hat', 'comments_hat', 'likes_hat']])

tmp = data_valid_test[['user_id', 'repost', 'comments', 'likes', 'repost_hat', 'comments_hat', 'likes_hat']]

m = data_train.groupby(['user_id'])[['repost', 'comments', 'likes']].median()
m = m.sort_values(by='repost', ascending=False)
m.columns = ['repost_hat', 'comments_hat', 'likes_hat']
tmp = pd.merge(tmp, m, how='left', on='user_id')

tmp['repost_hat_y'] = tmp['repost_hat_y'].fillna(tmp['repost_hat_x'])
tmp['comments_hat_y'] = tmp['comments_hat_y'].fillna(tmp['comments_hat_x'])
tmp['likes_hat_y'] = tmp['likes_hat_y'].fillna(tmp['likes_hat_x'])
tmp['repost_hat'], tmp['comments_hat'] , tmp['likes_hat']  =  tmp['repost_hat_y'] ,tmp['comments_hat_y'],tmp['likes_hat_y'] 
precision(tmp[['repost', 'comments', 'likes', 'repost_hat', 'comments_hat', 'likes_hat']])


### MSE on valid set

#mse(data_valid['repost'], data_valid['repost_hat'].apply(lambda x: int(x*1.3)))
print(mse(tmp['repost'], tmp['repost_hat']),\
         mse(tmp['comments'], tmp['comments_hat']), \
         mse(tmp['likes'], tmp['likes_hat']))

### 寻找最佳的整数转换系数

int_tmp = tmp.loc[:, ['repost', 'comments', 'likes', 'repost_hat', 'comments_hat', 'likes_hat']]
i = 1.0
best = [0, 0]
for j in range(1, 100):
    for col in ['repost_hat', 'comments_hat', 'likes_hat']:
        int_tmp.loc[:, col] = int_tmp.loc[:, col].apply(lambda x:int(i * x))
    score = precision(int_tmp)
    #print(i, score)
    if score > best[1]:
        best[0], best[1] = i, score
    else:
        pass
    i += 0.01
print(best)

### 复合模型

#读取和定义数据
data_predict=pd.read_csv('data_predict_tfidf.txt',index_col=[0],header=0)
x = data_predict[features_list]

data_predict['repost_hat'],         \
data_predict['comments_hat'],       \
data_predict['likes_hat']           \
        =                           \
        model_repost.predict(x),    \
        model_comments.predict(x),  \
        model_likes.predict(x)

tmp = data_predict[['user_id', 'weibo_id', 'repost_hat', 'comments_hat', 'likes_hat']]
m = data_train.groupby(['user_id'])[['repost', 'comments', 'likes']].median()
m = m.sort_values(by='repost', ascending=False)
m.columns = ['repost_hat', 'comments_hat', 'likes_hat']
tmp = pd.merge(tmp, m, how='left', on='user_id')

tmp['repost_hat_y']   = tmp['repost_hat_y'].fillna(tmp['repost_hat_x'])
tmp['comments_hat_y'] = tmp['comments_hat_y'].fillna(tmp['comments_hat_x'])
tmp['likes_hat_y']    = tmp['likes_hat_y'].fillna(tmp['likes_hat_x'])
tmp['repost_hat'], tmp['comments_hat'] , tmp['likes_hat']  = tmp['repost_hat_y'] ,tmp['comments_hat_y'],tmp['likes_hat_y'] 

if noise_cancellation == True:
    tmp[tmp['logit'] == 0]['repost_hat'], tmp[tmp['logit'] == 0]['comments_hat'] , tmp[tmp['logit'] == 0]['likes_hat']  = 0, 0, 0

result = tmp[['user_id', 'weibo_id', 'repost_hat', 'comments_hat', 'likes_hat']]

for col in ['repost_hat', 'comments_hat', 'likes_hat']:
    result[col] = result[col].apply(lambda x:int(1.3 * x)) # best mutiplier
result.head()

 # 按照线上提交要求重组并存储数据
result = result.assign(
    predict=result.repost_hat.astype(str) + ',' +
    result.comments_hat.astype(str)       + ',' +
    result.likes_hat.astype(str))         # longer but faster than apply lambda
result = result[['user_id', 'weibo_id', 'predict']]
result.columns = ['uid', 'mid', 'predict']
result.to_csv("./result.txt", header=False, sep="\t", index=False)

# Save the model

### SAVE THE MODEL TO DISK
joblib.dump(model_repost, './model/model_repost.sav')
joblib.dump(model_repost, './model/model_comments.sav')
joblib.dump(model_repost, './model/model_likes.sav')
